update from origin

+++ modules/library/coverage/src/main/java/org/geotools/coverage/grid/io/FileSystemFileSetManager.java
@@ -1,16 +1,31 @@
+/*
+ *    GeoTools - The Open Source Java GIS Toolkit
+ *    http://geotools.org
+ *
+ *    (C) 2013 - 2016, Open Source Geospatial Foundation (OSGeo)
+ *
+ *    This library is free software; you can redistribute it and/or
+ *    modify it under the terms of the GNU Lesser General Public
+ *    License as published by the Free Software Foundation;
+ *    version 2.1 of the License.
+ *
+ *    This library is distributed in the hope that it will be useful,
+ *    but WITHOUT ANY WARRANTY; without even the implied warranty of
+ *    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+ *    Lesser General Public License for more details.
+ */
 package org.geotools.coverage.grid.io;
 
 import java.io.File;
 import java.util.ArrayList;
 import java.util.Collections;
-import java.util.Iterator;
 import java.util.List;
 import java.util.logging.Level;
 import java.util.logging.Logger;
 
-public class FileSystemFileSetManager implements FileSetManager{
+public class FileSystemFileSetManager implements FileSetManager {
 
-    private static Logger LOGGER = Logger.getLogger(FileSystemFileSetManager.class.toString()); 
+    private static Logger LOGGER = Logger.getLogger(FileSystemFileSetManager.class.toString());
 
     private List<String> fileSet = Collections.synchronizedList(new ArrayList<String>());
 
@@ -38,15 +53,20 @@ public class FileSystemFileSetManager implements FileSetManager{
                 final File file = new File(filePath);
                 if (file.isDirectory()) {
                     File[] files = file.listFiles();
-                    for (File _file: files) {
-                        deleteFile(_file);
+                    if (files != null) {
+                        for (File _file : files) {
+                            deleteFile(_file);
+                        }
                     }
                 }
                 deleteFile(file);
-            } catch (Throwable t){
+            } catch (Throwable t) {
                 if (LOGGER.isLoggable(Level.FINE)) {
-                    LOGGER.fine("Exception occurred while deleting file: " + filePath + 
-                            "\n" + t.getLocalizedMessage());
+                    LOGGER.fine(
+                            "Exception occurred while deleting file: "
+                                    + filePath
+                                    + "\n"
+                                    + t.getLocalizedMessage());
                 }
             }
         }
@@ -65,11 +85,10 @@ public class FileSystemFileSetManager implements FileSetManager{
     @Override
     public void purge() {
         if (!fileSet.isEmpty()) {
-            String files[] = (String[])fileSet.toArray(new String[fileSet.size()]);
-            for (String filePath: files) {
+            String[] files = (String[]) fileSet.toArray(new String[fileSet.size()]);
+            for (String filePath : files) {
                 removeFile(filePath);
             }
         }
     }
-
 }
